tests/wpt/web-platform-tests/html/semantics/popovers/popover-defaultopen.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel=author href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div popover id=p1 defaultopen>This is a popover, which should be open upon load</div>
<script></script> <!-- Possibly yield the parser, just to double-check -->
<div popover id=p2 defaultopen>This is a second popover with defaultopen, which should NOT be open upon load</div>
<div popover id=p3>Also not visible</div>

<div popover=manual id=p4 defaultopen>This is a manual popover with defaultopen, which should be open upon load</div>
<div popover=manual id=p5 defaultopen>This is a manual popover with defaultopen, which should be open upon load</div>

<script>
  requestAnimationFrame(() => {
    requestAnimationFrame(() => {
      test(function(){
        assert_true(p1.matches(':open'),'defaultopen should open the popover');
        assert_true(p1.hasAttribute('defaultopen'));
        assert_true(p1.defaultOpen,'defaultopen should be reflected in the IDL attribute');
        assert_false(p2.matches(':open'), 'Only the first popover with defaultopen should be open on load');
        assert_true(p2.hasAttribute('defaultopen'),'defaultopen should be present/true, even if not opened');
        assert_true(p2.defaultOpen,'defaultopen should be present/true, even if not opened');

        assert_true(p4.matches(':open'),'defaultopen should open all manual popovers');
        assert_true(p5.matches(':open'),'defaultopen should open all manual popovers');

        assert_false(p3.matches(':open'));
        p3.setAttribute('defaultopen','');
        assert_false(p3.matches(':open'), 'Changing defaultopen should not affect open status');
        assert_true(p3.hasAttribute('defaultopen'));
        assert_true(p3.defaultOpen,'defaultopen should still reflect to IDL');

        p1.removeAttribute('defaultopen');
        assert_true(p1.matches(':open'),'removing defaultopen should not close the popover');
        assert_false(p1.hasAttribute('defaultopen'),'...but it should reflect to IDL');

        p1.hidePopover();
      }, "The defaultopen attribute should affect page load only");
    });
  });
</script>
